Adding basic physics and stubbing out classes

+++ expotek_mn/src/main/java/com/expotek/mn/basic/Acceleration.java
@@ -2,13 +2,13 @@ package com.expotek.mn.basic;
 
 public class Acceleration {
 
-    private int ms2;
+    private double ms2;
 
-    public int getMs2() {
+    public double getMs2() {
         return ms2;
     }
 
-    public void setMs2(int ms2) {
+    public void setMs2(double ms2) {
         this.ms2 = ms2;
     }
 }
